pihaalpr/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlmodel==0.0.22
pydantic==2.10.4
pydantic-settings==2.7.0
aiohttp==3.11.11
httpx==0.28.1
paho-mqtt==2.1.0
apscheduler==3.10.4
python-multipart==0.0.20
numpy
Pillow
Pillow
